State the resolution rule conditionally and split the properties
The guide claimed callers always pass resolved program paths; the rule
is conditional — only callers that must not let the injected PATH
select the executable need one — and the probe test's doc comment said
the opposite of the truth (Unix lookup uses the child's PATH once one
is set, not the parent's). Align both statements. Build the with_path
doctest's example value with join_paths so Windows readers do not copy
a Unix-only literal, and split the property module into a sibling file
via #[path] to bring env_path_tests.rs back under the 400-line ceiling.
